added main.py and changed the urls
