+++ flaskr/__init__.py
@@ -34,5 +34,8 @@ def create_app(test_config=None):
 
     from . import auth
     app.register_blueprint(auth.bp)
-    
+
+    from . import main
+    app.register_blueprint(main.bp)
+
     return app
