+++ flaskr/auth.py
@@ -1,7 +1,7 @@
 import functools
 
 from flask import (
-    Blueprint, flash, g, redirect, render_template, request, session, url_for
+    Blueprint, flash, g, redirect, render_template, request, session, url_for, __main__
 )
 from werkzeug.security import check_password_hash, generate_password_hash
 
@@ -38,7 +38,7 @@ def register():
                 ).fetchone()
                 session.clear()
                 session['user_id'] = user['id']
-                return redirect(url_for('auth.create_bio'))
+                return redirect(url_for('main.create_bio'))
             flash(error)
         #else if the login button is clicked load the login inputs
         else:
@@ -56,7 +56,7 @@ def register():
             if error is None:
                 session.clear()
                 session['user_id'] = user['id']
-                return redirect(url_for('auth.home'))
+                return redirect(url_for('main.home'))
  
             flash(error)
 
@@ -76,175 +76,7 @@ def load_logged_in_user():
 def logout():
     session.clear()
     return redirect(url_for('auth.start_page'))
-    
-	
-@bp.route('/create_bio')
-def create_bio():
-    user_id = session.get('user_id')
-    db = get_db()
-    if user_id is None:
-        return redirect(url_for('auth.start_page'))
-    user = db.execute(
-        'SELECT * FROM user WHERE id = ?', (user_id,)
-    ).fetchone()
-    if user['first'] is None:
-        first = ""
-    else:
-        first = user['first']
-    if user['email'] is None:
-        email = ""
-    else:
-        email = user['email']
-    if user['last'] is None:
-        last = ""
-    else:
-        last = user['last']
-    if user['address_line1'] is None:
-        address_line1 = ""
-    else:
-        address_line1 = user['address_line1']
-    if user['address_line2'] is None:
-        address_line2 = ""
-    else:
-        address_line2 = user['address_line2']
-    if user['username'] is None:
-        username = ""
-    else:
-        username = user['username']
-    user_details = {      
-        'first': first,
-        'last': last,
-        'email': email,
-        'address1': address_line1,
-        'address2': address_line2,
-        'username': username   
-    }
-    return render_template('main/create_bio.html', user=user_details)
-   
-@bp.route('/bio')
-def bio():
-    user_id = session.get('user_id')
-    db = get_db()
-    if user_id is None:
-        return redirect(url_for('auth.start_page'))
-    user = db.execute(
-        'SELECT * FROM user WHERE id = ?', (user_id,)
-    ).fetchone()
-    if user['first'] is None:
-        first = ""
-    else:
-        first = user['first']
-    if user['email'] is None:
-        email = ""
-    else:
-        email = user['email']
-    if user['last'] is None:
-        last = ""
-    else:
-        last = user['last']
-    if user['address_line1'] is None:
-        address_line1 = ""
-    else:
-        address_line1 = user['address_line1']
-    if user['address_line2'] is None:
-        address_line2 = ""
-    else:
-        address_line2 = user['address_line2']
-    if user['username'] is None:
-        username = ""
-    else:
-        username = user['username']
-    user_details = {      
-        'first': first,
-        'last': last,
-        'email': email,
-        'address1': address_line1,
-        'address2': address_line2,
-        'username': username   
-    }
-    return render_template('main/bio.html', user=user_details)
-   
-@bp.route('/edit_bio')
-def edit_bio():
-    user_id = session.get('user_id')
-    db = get_db()
-    if user_id is None:
-        return redirect(url_for('auth.start_page'))
-    user = db.execute(
-        'SELECT * FROM user WHERE id = ?', (user_id,)
-    ).fetchone()
-    if user['first'] is None:
-        first = ""
-    else:
-        first = user['first']
-    if user['email'] is None:
-        email = ""
-    else:
-        email = user['email']
-    if user['last'] is None:
-        last = ""
-    else:
-        last = user['last']
-    if user['address_line1'] is None:
-        address_line1 = ""
-    else:
-        address_line1 = user['address_line1']
-    if user['address_line2'] is None:
-        address_line2 = ""
-    else:
-        address_line2 = user['address_line2']
-    if user['username'] is None:
-        username = ""
-    else:
-        username = user['username']
-    user_details = {      
-        'first': first,
-        'last': last,
-        'email': email,
-        'address1': address_line1,
-        'address2': address_line2,
-        'username': username   
-    }
-    return render_template('main/edit_bio.html', user=user_details)
-    
-@bp.route('/edit_bio', methods=('GET', 'POST'))
-def edit_bio_form():
-    user_id = session.get('user_id')
-    if request.method == 'POST':
-        db = get_db()
-        error = None
-        #if the register button is clicked load the register inputs
-        first = request.form['first']
-        last = request.form['last']
-        email = request.form['email']
-        username = request.form['username']
-        address1 = request.form['address1']
-        address2 = request.form['address2']
-        sql = 'UPDATE user SET first = ?, last = ?, email = ?, username = ?, address_line1 = ?, address_line2 = ? WHERE id = ?'
-        val = (first, last, email, username, address1, address2, user_id)
-        db.execute(sql, val)
-        db.commit()
-        return redirect(url_for('auth.db'))
-    
-@bp.route('/create_bio', methods=('GET', 'POST'))
-def create_bio_form():
-    user_id = session.get('user_id')
-    if request.method == 'POST':
-        db = get_db()
-        error = None
-        #if the register button is clicked load the register inputs
-        first = request.form['first']
-        last = request.form['last']
-        email = request.form['email']
-        username = request.form['username']
-        address1 = request.form['address1']
-        address2 = request.form['address2']
-        sql = 'UPDATE user SET first = ?, last = ?, email = ?, username = ?, address_line1 = ?, address_line2 = ? WHERE id = ?'
-        val = (first, last, email, username, address1, address2, user_id)
-        db.execute(sql, val)
-        db.commit()
-        return redirect(url_for('auth.db'))
-       
+
 
 @bp.route('/db')
 def db():
@@ -272,26 +104,12 @@ def update_email():
         db.commit()
         return redirect(url_for('auth.db'))
 
-@bp.route('/home')
-def home():
-    user_id = session.get('user_id')
-    db = get_db()
-    if user_id is None:
-        return redirect(url_for('auth.start_page'))
-    user = db.execute(
-        'SELECT * FROM user WHERE id = ?', (user_id,)
-    ).fetchone()
-    user_details = {      
-		'email': user['email'],
-    }
-    return render_template('main/home.html', user=user_details)
-    
     
 @bp.route('/start_page')
 def start_page():
     user_id = session.get('user_id')
     if user_id is not None:
-        return redirect(url_for('main.create_bio'))
+        return redirect(url_for('main.home'))
     else:
         return render_template('auth/start_page.html')
