+++ flaskr/templates/main/bio.html
@@ -58,7 +58,7 @@
 							<h3>{{user.address2}}</h3> 
 						</div>
 					</div>
-                    <a href="{{url_for('auth.edit_bio')}}"><button class="btn waves-effect waves-light">Edit</button></a>
+                    <a href="{{url_for('main.edit_bio')}}"><button class="btn waves-effect waves-light">Edit</button></a>
 			</div>
 		</div>
     <div class="parallax red"></div>
