+++ flaskr/templates/main/db.html
@@ -71,9 +71,9 @@
 <button class="btn waves-effect waves-light" type="submit" name="button">new email</button>
 </form>
 	<a href="{{ url_for('auth.logout') }}">log out</a>
-    <a href="{{ url_for('auth.bio') }}">View Bio</a>
-    <a href="{{ url_for('auth.create_bio') }}">Create Bio</a>
-    <a href="{{url_for('auth.home')}}">Home</a>
+    <a href="{{ url_for('main.bio') }}">View Bio</a>
+    <a href="{{ url_for('main.create_bio') }}">Create Bio</a>
+    <a href="{{url_for('main.home')}}">Home</a>
         
 <!-- Footer Section -->
 	<footer class="page-footer white">
